fix(*): type cleanup
