--- packages/portal/spec-renderer/sandbox/spec-details-sandbox/index.ts
@@ -1,6 +1,0 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-
-const app = createApp(App)
-
-app.mount('#app')
