--- packages/portal/spec-renderer/sandbox/spec-operations-list-sandbox/index.ts
@@ -1,7 +1,0 @@
-import { createApp } from 'vue'
-import '@kong/kongponents/dist/style.css'
-import App from './App.vue'
-
-const app = createApp(App)
-
-app.mount('#app')
